Refactor http server so that it can be started in a test
