+++ pkg/server/admin_http.go
@@ -1,4 +1,4 @@
-package main
+package server
 
 import (
 	"errors"
@@ -11,14 +11,14 @@ import (
 	"github.com/labstack/echo/v4/middleware"
 )
 
-func initAdminRouter(e *echo.Echo, db storage.Driver, config Config) {
+func initAdminRouter(e *echo.Echo, db storage.Driver, adminAPIKey string) {
 	handler := &AdminRoutesHandler{
 		db: db,
 	}
 	adminRoutes := e.Group("/admin")
 	adminRoutes.Use(middleware.KeyAuth(func(key string, c echo.Context) (bool, error) {
 		// Probably not ok: this isn't a constant time compare
-		return key == config.Auth.Admin.ApiKey, nil
+		return key == adminAPIKey, nil
 	}))
 	adminRoutes.POST("/jobs", handler.AddJob)
 	adminRoutes.GET("/jobs/:job_id", handler.GetJob)
